% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegionSetDB-class.R
\docType{methods}
\name{addRegionSet,RegionSetDB-method}
\alias{addRegionSet,RegionSetDB-method}
\alias{addRegionSet}
\title{addRegionSet-methods}
\usage{
\S4method{addRegionSet}{RegionSetDB}(.object, pid, name, description, assembly,
  collection, collection_description, source, parser, file)
}
\arguments{
\item{.object}{\code{\linkS4class{RegionSetDB}} object}

\item{pid}{a unique region set id used for the parser (character)}

\item{name}{a name for the region set (character)}

\item{description}{a brief description of the region set (character)}

\item{assembly}{the genome assembly of the region set (character)}

\item{collection}{a collection or group this region set belongs to (character)}

\item{collection_description}{a brief description of the collection/group (character)}

\item{source}{a reference or source from where the region set was obtained (character)}

\item{parser}{name of the parser function used to import the region set (character)}

\item{file}{the filename or input argument to parser function (character)}
}
\value{
a modified \code{\linkS4class{RegionSetDB}} object with the region set added
}
\description{
parses a region set and adds it to the \code{\linkS4class{RegionSetDB}} object
}
\author{
Fabian Mueller
}
