% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegionSetDB-class.R
\docType{methods}
\name{addRegionSetFromDataFrame,RegionSetDB-method}
\alias{addRegionSetFromDataFrame,RegionSetDB-method}
\alias{addRegionSetFromDataFrame}
\title{addRegionSetFromDataFrameFromDataFrame-methods}
\usage{
\S4method{addRegionSetFromDataFrame}{RegionSetDB}(.object, df)
}
\arguments{
\item{.object}{\code{\linkS4class{RegionSetDB}} object}

\item{df}{a data.frame object containing the required columns (see \code{\link{addRegionSet,RegionSetDB-metho}} for a brief description of the required columns)}
}
\value{
a modified \code{\linkS4class{RegionSetDB}} object with the region sets added
}
\description{
parses multiple region sets and adds them to the \code{\linkS4class{RegionSetDB}} object. The parsing parameters are read from a data.frame
}
\author{
Fabian Mueller
}
