% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegionSetDB-class.R
\docType{methods}
\name{collections,RegionSetDB-method}
\alias{collections,RegionSetDB-method}
\alias{collections}
\title{collections-methods}
\usage{
\S4method{collections}{RegionSetDB}(.object, genomeNames = genomes(.object))
}
\arguments{
\item{.object}{\code{\linkS4class{RegionSetDB}} object}

\item{genomeNames}{optional vector of genome assemblies for which the collection should be retrieved}
}
\value{
Character vector of collection names contained in the dataset
}
\description{
Return collectionss contained in the dataset
}
\author{
Fabian Mueller
}
