% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegionSetDB-class.R
\docType{methods}
\name{dbDir,RegionSetDB-method}
\alias{dbDir,RegionSetDB-method}
\alias{dbDir}
\title{dbDir-methods}
\usage{
\S4method{dbDir}{RegionSetDB}(.object)
}
\arguments{
\item{.object}{\code{\linkS4class{RegionSetDB}} object}
}
\description{
Retrieve or set the directories were the object and region sets are stored
}
\author{
Fabian Mueller
}
