% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/export.R
\docType{methods}
\name{export2Lola,RegionSetDB-method}
\alias{export2Lola,RegionSetDB-method}
\alias{export2Lola}
\title{export2Lola-methods}
\usage{
\S4method{export2Lola}{RegionSetDB}(.object, destDir, doIndex = TRUE)
}
\arguments{
\item{.object}{\code{\linkS4class{RegionSetDB}} object}

\item{destDir}{destination directory}

\item{doIndex}{index the LOLA database for faster re-loading}
}
\value{
nothing of particular interest
}
\description{
Export a \code{\linkS4class{RegionSetDB}} object as LOLA database
}
\author{
Fabian Mueller
}
