% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegionSetDB-class.R
\docType{methods}
\name{genomes,RegionSetDB-method}
\alias{genomes,RegionSetDB-method}
\alias{genomes}
\title{genomes-methods}
\usage{
\S4method{genomes}{RegionSetDB}(.object)
}
\arguments{
\item{.object}{\code{\linkS4class{RegionSetDB}} object}
}
\value{
Character vector of genome assemblies contained in the dataset
}
\description{
Return genome assemblies contained in the dataset
}
\author{
Fabian Mueller
}
