% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegionSetDB-class.R
\docType{methods}
\name{length,RegionSetDB-method}
\alias{length,RegionSetDB-method}
\title{length-methods}
\usage{
\S4method{length}{RegionSetDB}(x)
}
\arguments{
\item{x}{\code{\linkS4class{RegionSetDB}} object}
}
\value{
integer indicating the number of region sets in the DB
}
\description{
Return the number of region sets in the DB
}
\author{
Fabian Mueller
}
