% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegionSetDB-class.R
\name{loadRegionSetDB}
\alias{loadRegionSetDB}
\title{loadRegionSetDB}
\usage{
loadRegionSetDB(path, loadRegionSets = FALSE)
}
\arguments{
\item{path}{Location of saved \code{\linkS4class{RegionSetDB}} object}

\item{loadRegionSets}{should the individual region sets be loaded into the object. If not this will be done as they are required in a lazy-load fashion}
}
\value{
\code{\linkS4class{RegionSetDB}} object
}
\description{
Load a RegionSetDB dataset from disk
}
\author{
Fabian Mueller
}
