% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegionSetDB-class.R
\docType{class}
\name{RegionSetDB-class}
\alias{RegionSetDB-class}
\alias{RegionSetDB}
\alias{intialize,RegionSetDB-method}
\title{RegionSetDB}
\usage{
RegionSetDB(parserAnnot = NULL)
}
\arguments{
\item{parserAnnot}{A data frame containing annotations for the region sets to be parsed.
requires the following columns: \code{"pid", "name", "description", "assembly", "collection", "collection_description", "source", "parser", "file"}}
}
\description{
A class for storing a database of genomic region annotations
}
\section{Slots}{

\describe{
  \item{\code{parserSummary}}{
	A summary table containing information for parsing region sets
  }
  \item{\code{sampleAnnot}}{
	Sample annotation Table
  }
  \item{\code{genomes}}{
	Character vector of genome assemblies contained in the database
  }
}
}

\section{Methods}{

\describe{
   \item{\code{\link{blubb,RegionSetDB-method}}}{
     some function
   }
}
}

\author{
Fabian Mueller
}
