% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegionSetDB-class.R
\docType{methods}
\name{regionSetNames,RegionSetDB-method}
\alias{regionSetNames,RegionSetDB-method}
\alias{regionSetNames}
\title{regionSetNames-methods}
\usage{
\S4method{regionSetNames}{RegionSetDB}(
  .object,
  genomeNames = genomes(.object),
  collectionNames = collections(.object)
)
}
\arguments{
\item{.object}{\code{\linkS4class{RegionSetDB}} object}

\item{genomeNames}{optional vector of genome assemblies for which the names should be retrieved}

\item{collectionNames}{optional vector of collections for which the names should be retrieved}
}
\value{
Character vector of region set names contained in the dataset
}
\description{
Return region set names contained in the dataset
}
\author{
Fabian Mueller
}
