% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RegionSetDB-class.R
\name{saveRegionSetDB}
\alias{saveRegionSetDB}
\title{saveRegionSetDB}
\usage{
saveRegionSetDB(.object, path = dbDir(.object))
}
\arguments{
\item{.object}{\code{\linkS4class{RegionSetDB}} object}

\item{path}{destination to save the object to}
}
\value{
nothing of particular interest
}
\description{
Save a RegionSetDB dataset to disk for later loading
}
\author{
Fabian Mueller
}
